OPENEUROPA-1292: Bring test inline with common Drupal development practices.
- Browser tests are functional tests.
- Do not split code on multiple lines when it doesn't benefit readability.
- Improve documentation.
